+++ BlackJack/Card.h
@@ -8,20 +8,20 @@ namespace BlackJack
 
 	class Card
 	{
-	public:
-		Card();
-		Card(const Card&);
-		Card& operator=(const Card&);
-		Card(CardNames, SuiteNames);
-		const CardNames GetName() const;
-		const SuiteNames GetSuite() const;
-		const short unsigned int GetHighValue() const;
-		const short unsigned int GetLowValue() const;
-	private:
-		CardNames name;
-		SuiteNames suite;
-		short unsigned int highvalue;
-		short unsigned int lowvalue;
+		public:
+			Card();
+			Card(const Card&);
+			Card& operator=(const Card&);
+			Card(CardNames, SuiteNames);
+			const CardNames GetName() const;
+			const SuiteNames GetSuite() const;
+			const uint8_t GetHighValue() const;
+			const uint8_t GetLowValue() const;
+		private:
+			CardNames name;
+			SuiteNames suite;
+			uint8_t highvalue;
+			uint8_t lowvalue;
 	};
 
 }
